% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{Msfrags}
\format{
A character vector of length 4 with peptide sequences
}
\source{
data-raw/data.R
}
\usage{
MSfrags
}
\description{
Some fragments with protein modifications that we might
expect to find from the CCR5 gene in a mass spec experiment
}
\keyword{datasets}
